requirements.txt
arrow==0.17.0
asgiref==3.3.1
blessed==1.17.12
cairocffi==1.2.0
CairoSVG==2.5.1
cffi==1.14.4
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.5
django-picklefield==3.0.1
django-q==1.3.4
html5lib==1.1
Pillow==8.1.0
pycparser==2.20
Pyphen==0.10.0
python-dateutil==2.8.1
pytz==2020.5
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
